consistent wip through test refactoring
